Cursor/debug trait resolution and complete add operator tests c181 (#42)
* Checkpoint before follow-up message

* Checkpoint before follow-up message

* Implement trait-based type checking and runtime evaluation for binary ops


* Implement polymorphic type inference for binary operators with traits


* Limit constraint resolution to arithmetic operators


* Improve Add trait handling and primitive type operations


* Checkpoint before follow-up message

* Add support for distinguishing numeric literals with original token


* Checkpoint before follow-up message

* Implement numeric operators with trait-based polymorphic support


* Implement safe division and unified math trait system with tests


* Remove Int type, unify numeric types to Float


* Fix tests after killing Int

* implement add for list

---------

+++ src/lexer/__tests__/lexer.test.ts
@@ -144,7 +144,7 @@ describe('Lexer', () => {
 		});
 
 		test('should recognize primitive type keywords', () => {
-			const primitives = ['Int', 'Number', 'String', 'Unit', 'List'];
+			const primitives = ['Float', 'Number', 'String', 'Unit', 'List'];
 
 			for (const primitive of primitives) {
 				const tokens = getTokenValues(primitive);
